Move some tests out of tests/ui
