+++ tests/ui/attributes/no_link/auxiliary/empty-crate-1.rs
@@ -1,3 +1,1 @@
-//@ no-prefer-dynamic
-
 #![crate_type = "dylib"]
