+++ tests/ui/attributes/no_link/auxiliary/empty-crate-2.rs
