+++ tests/ui/cross-crate/auxiliary/inline-cross-crate.rs
