+++ tests/ui/macros/auxiliary/fancy-panic.rs
