--- tests/ui/no-link-unknown-crate.rs
@@ -1,4 +1,0 @@
-#[no_link]
-extern crate doesnt_exist; //~ ERROR can't find crate
-
-fn main() {}
