+++ tests/ui/pattern/auxiliary/cross-crate-enum-pattern.rs
