+++ tests/ui/privacy/auxiliary/impl-privacy-cross-crate-1.rs
