+++ tests/ui/privacy/auxiliary/impl-privacy-cross-crate-2.rs
